Extracted shared markdown util
